dom/indexedDB/test/test_put_get_values.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const name = window.location.pathname;
      const description = "My Test Database";
      const objectStoreName = "Objects";

      let testString = { key: 0, value: "testString" };
      let testInt = { key: 1, value: 1002 };

      let request = mozIndexedDB.open(name, 1, description);
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      let event = yield;

      let db = event.target.result;

      let objectStore = db.createObjectStore(objectStoreName,
                                             { autoIncrement: 0 });

      request = objectStore.add(testString.value, testString.key);
      request.onerror = errorHandler;
      request.onsuccess = function(event) {
        is(event.target.result, testString.key, "Got the right key");
        request = objectStore.get(testString.key);
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          is(event.target.result, testString.value, "Got the right value");
        };
      };

      request = objectStore.add(testInt.value, testInt.key);
      request.onerror = errorHandler;
      request.onsuccess = function(event) {
        is(event.target.result, testInt.key, "Got the right key");
        request = objectStore.get(testInt.key);
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          is(event.target.result, testInt.value, "Got the right value");
          finishTest();
        };
      }

      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
